Combat Test Feedback Implemented
Resolved the bugs from the build, including the downed state not entering and the projectile speeds not resetting. I implemented the feedback from combat testing on the existing attacks to make them more challenging but more fair (for example, some tracking attacks now always slightly miss the player on the first shot so they can react). Tweaked energy values to ensure that the enemy stays in the downed state longer (should be enough for 6-7 attacks from the player). Plus some infrastructure stuff in the Spawner logic to accommodate the new attacks I'm working on.